color ghosting

+++ drivers/hdmi/hdmi.h
@@ -32,7 +32,7 @@ extern "C" {
 #define TEXTMODE_COLS 53
 #define TEXTMODE_ROWS 30
 
-#define RGB888(r, g, b) ((r<<16) | (g << 8 ) | b )
+#define RGB888(r, g, b) ( ((uint32_t)(r) << 16) | ((uint32_t)(g) << 8 ) | (b) )
 
 // TODO: Сделать настраиваемо
 static const uint8_t textmode_palette[16] = {
